Fixed significant interaction
As we adjusted the mcols in the significant itneraciton file to get the sigificance and raw contact probailities for our thesis table. but this throws off the Gviz graphics hence we un did it for the app. The interacitons are the same but the probabilities and Q values are not assoicated with them